[BugFix] Categorical spec samples the right dtype when masked (#2980)

+++ torchrl/data/tensor_specs.py
@@ -3926,7 +3926,7 @@ class Categorical(TensorSpec):
                 "The last dimension of the mask must match the number of action allowed by the "
                 f"Categorical spec. Got mask.shape={self.mask.shape} and n={n}."
             )
-        out = torch.multinomial(mask_flat.float(), 1).reshape(shape_out)
+        out = torch.multinomial(mask_flat.float(), 1).reshape(shape_out).to(self.dtype)
         return out
 
     def index(
